Local Repositories/Apps/Pathways Companion/backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
supabase==2.3.4
python-dotenv==1.0.0
